refactor(api): deprecate POST schooling-registration-dependent-users/external-user-token
in favor of POST sco-organization-learners/external
