+++ api/lib/application/organization-learner-dependent-users/index.js
@@ -13,36 +13,6 @@ exports.register = async function (server) {
   server.route([
     {
       method: 'POST',
-      path: '/api/schooling-registration-dependent-users/external-user-token',
-      config: {
-        auth: false,
-        handler: organizationLearnerDependentUserController.createUserAndReconcileToOrganizationLearnerFromExternalUser,
-        validate: {
-          options: {
-            allowUnknown: false,
-          },
-          payload: Joi.object({
-            data: {
-              attributes: {
-                'campaign-code': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                'external-user-token': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                birthdate: Joi.date().format('YYYY-MM-DD').raw().required(),
-                'access-token': Joi.string().allow(null).optional(),
-              },
-              type: 'external-users',
-            },
-          }),
-        },
-        notes: [
-          "Cette route crée un compte utilisateur suite à une connexion provenant d'un IDP externe (GAR). " +
-            "Les informations sont fournies dans un token. Elle réconcilie également cet utilisateur avec l'inscription " +
-            "de l'élève au sein de l'organisation qui a créé la campagne.",
-        ],
-        tags: ['api', 'organizationLearnerDependentUser'],
-      },
-    },
-    {
-      method: 'POST',
       path: '/api/schooling-registration-dependent-users/password-update',
       config: {
         pre: [
