+++ api/lib/infrastructure/serializers/jsonapi/sco-organization-learner-serializer.js
@@ -12,4 +12,10 @@ module.exports = {
       attributes: ['lastName', 'firstName', 'birthdate', 'username'],
     }).serialize(scoOrganizationLearner);
   },
+
+  serializeExternal(scoOrganizationLearner) {
+    return new Serializer('external-users', {
+      attributes: ['accessToken'],
+    }).serialize(scoOrganizationLearner);
+  },
 };
