+++ api/tests/unit/infrastructure/serializers/jsonapi/sco-organization-learner-serializer_test.js
@@ -62,4 +62,28 @@ describe('Unit | Serializer | JSONAPI | sco-organization-learner-serializer', fu
       expect(json).to.deep.equal(expectedSerializedOrganizationLearner);
     });
   });
+
+  describe('#serializeExternal', function () {
+    it('should convert into JSON API data', function () {
+      // given
+      const organizationLearner = {
+        accessToken: 'some token',
+      };
+
+      const expectedSerializedOrganizationLearner = {
+        data: {
+          attributes: {
+            'access-token': organizationLearner.accessToken,
+          },
+          type: 'external-users',
+        },
+      };
+
+      // when
+      const json = serializer.serializeExternal(organizationLearner);
+
+      // then
+      expect(json).to.deep.equal(expectedSerializedOrganizationLearner);
+    });
+  });
 });
